wanted long instead of int, logic was correct

+++ src/Program.java
@@ -401,8 +401,8 @@ public class Program {
 
 //    Take an array of ints and return the minimum and maximum sums from the numbers provided
     static void miniMaxSum(int[] arr) {
-        int min = 0;
-        int max = 0;
+        long min = 0;
+        long max = 0;
         Arrays.sort(arr);
 
         for (int i = 0; i < arr.length - 1; i++) {
